Remove the duplicated Front-End folder and merge the changes in there into the one Static folder.
Please do not duplicate the files like this, let's all just work on the one copy.

This change also modifies how the rent and return tabs function, instead of three identical blocks we now have one for rent and one for return. The rent and return option shows both, on submit we can simply grab the values from the blocks we need.