Implement search button functionality.
When submit is pressed, we check the validity of the input and send a et request to flask with the station name, date and time, and whether this is a rent or return query. We check date validity such date if the timepoint is before now, we set the selected time to now.

Add a specific status 500 error handler in case no stations matching the given name are found.

Add a global var in our js file for the currently open tab. Since multiple parts of our layout depend on this, I'm sure this will prove useful.